Various scalar type mismatches

+++ src/PhpSpec/Formatter/Presenter/Exception/GenericPhpSpecExceptionPresenter.php
@@ -30,7 +30,7 @@ final class GenericPhpSpecExceptionPresenter extends AbstractPhpSpecExceptionPre
     protected function presentFileCode(string $file, int $lineno, int $context = 6): string
     {
         $lines  = explode(PHP_EOL, file_get_contents($file));
-        $offset = max(0, $lineno - ceil($context / 2));
+        $offset = (int)max(0, $lineno - ceil($context / 2));
         $lines  = \array_slice($lines, $offset, $context);
 
         $text = PHP_EOL;
